requirements.txt
psycopg[binary]>=3.2,<4
httpx>=0.27,<1
python-dotenv>=1.0,<2
duckduckgo-search>=6.0,<8
